requirements.txt
fastapi
geocoder
gunicorn
pydantic
uvicorn
uvloop
httptools
firebase_admin
google-cloud-logging